Implement system skinning, remove pgeu data
System skinning allows a "merged view" of the core code with that of a
"skin". A skin represents the local deployment, such as "PostgreSQL
Europe". Individual conferences are still skinned individually on top
of this.

As part of this, remove all PGEU website specific content, as that is
now moved to the PGEU specific skin. Also removes the cmutuel
application module as nobody outside PGEU needs this.

Update the global templates to remove mentions of PostgreSQL Europe,
and let PGEU override those templates as necessary. Front page
template just becomes an empty placeholder, and is expected to always
be overridden.

Makes several more modules configurable by setting ENABLE_NEWS,
ENABLE_MEMBERSHIP and ENABLE_ELECTIONS (we already had
ENABLE_BRAINTREE).

And finally, update the devsetup to be a bit more correct and also to
set up an uwsgi ini file to use for local testing.

Still lacking documentation and a few small parts are still PGEU
hardcoded, so some follow-up patches to come.