Add generic backend form/list support, add membership module
Fix some of the minor parts of the backend views/lists that did not work
outside of confreg.

As a demo of it working, add administration views for the membership
module, specifically allow editing of members and meetings.

Create a generic admin dashboard page on /admin/. This one links through
to the /events/admin/ page like before, and also to the new membership
module if the user has permissions.