Refactor and rename invoice PDF builders
Split apart the settings for Invoice PDF and Refund PDF builders, as
they can theoretically be loaded from different places.

Remove all PGEU specific contents from the invoices. This will now
instead go in a class that lives in the skin, if there is one. If there
is no skin, an "anonymous" invoice with basically no info will be
generated. Enough for testing, but invoices should really always be
skinned.

Rename invoices to BaseInvoice and BaseRefund to make this clear.

Make the invoice properly handle both VAT and non-VAT cases, so that it
can be used in skinned versions of both.

BREAKING change - this changes the way the invoices are configured and
called, so all skinned installations must update to cover this.