Move invoice management to the new layout/framework
This doesn't use the backend editors etc because it's a bit too complex
for that, but now renders in that framework instead of the frontend one.
Also gives it a visual refresh.

Does not touch the refund part as that's scheduled for replacement
anyway.

Should not affect any auto-genereated invoices, other than viewing them.