Re-think how refunds are handled
Basically reverse the flow - instead of having the invoice system
refund an invoice and trigger an action in the attached systems (such
as confreg or membership) to remove something, have the actual end
systems like confreg handle the cancelation and just tell the invoice
system what the refund is.

This makes the system a lot more flexible and makes it possible to
handle things like multi-attendee invoices that get partially refunded
more than once.

Other than the order-reversal:

1. Each invoice can now have multiple refund notes. In doing this, the
   view of "refunded invoices" has also been retired, as it makes less
   sense now.

2. Refund notices now include a "previously refunded" section if they
   are anything except the first refund note, to track across multiple
   refunds. This creates a breaking change in the API for PDF
   generation, but recent other updates to the skinning broke them all
   anyway.

3. For confreg, implement "refund patterns". This is a set of rules
   with "refund x% minus fees of y" rules that can be applied when
   canceling a registration. They are applied to the individual costs
   being refunded. They can also be given a start/end date during
   which they should be used, which is used as a hint in the UI.