Fix fieldsets for admin of conference registration entry
dietary needs field was misspelled, would cause a crash if editing a
conference that did not ask for dietary needs.

shareemail field was not in a fieldset, would cause a crash for editing
anything at all on a conference with email sharing enabled (clearly
that's not used much these days).