Add support for dynamic forms for json fields
This makes it possible to have some regluar form fields render into a
json field instead of into separate columns, within the framework of the
backend forms. This can be used to replace (some of) the fields that
requires manual json input (as long as the expected structure is simple)
with proper form fields.