Update migrations for python 3
All models were updated to have things like help text and verbose name
be unicode strings. This needs to be backfilled into the migrations as
well, otherwise upcoming migratinos will contain no-change changes.

For special reg types, in python3 the ordering coming out of the dict
can be random, thus generating new migrations all the time. So pre-sort
it to make it predictable. Doesn't make any actual difference in the
end, but it should make the migrations happier. Doesn't make any actual
difference in the end, but it should make the migrations happier.