Make payment methods configurable on a per-conference basis
We could previously select paymentm methods for the different
sponsorship levels, but there was a global configuration for payment
methods for a conference. This now moves into a per-conference setting
(configurable by superuser only). All existing conferences are populated
with whatever the current global value for auto is and then the auto
field is removed.