Fix trustly withdrawal fetching
Zero is not more than one. This clearly never worked, on account of us
not being able to test refunds on Trustly testing system due to a vendor
bug.Fix trustly withdrawal fetching

Zero is not more than one. This clearly never worked, on account of us
not being able to test refunds on Trustly testing system due to a vendor
bug.