Fix reporting of errors of API refund calls
Previously the success or error of an API refund would not bubble up,
and thus always be reported as "issued refund" even if it failed. Only
the report was incorrect, so confusingly enough the attempts would
continue. Fix this by reporting that it failed *and* that it will keep
retrying.