Attempt to deal with Trustly refunds in the wrong currency
When a Trustly transaction is refunded in a non-default currency the
actual refund is listed in that currency and not the primary one, which
makes it not match up in the accounting.

To deal with that, fetch the ledger rows for the day of the refund and
try to locate the actual amount there. If found, use this amount to
calculate "fake fees" (as the amount normally doesn't match perfecly due
to currency rates changing between payment and refund). This should
hopefully lead to a balanced accounting entry.