Handle Trustly payments in pending state
Trustly payments can remaining pending state for quite a long time, due
to banks being closed over weekends etc. In this case, bad things
happened when the invoice was automatically canceled.

However, unlike iban payments we know that the likelihood of the payment
actually being completed is very high. So instead of giving up, have the
system automatically extend the invoice cancel time when this happens.
This is done in two stages:

1. Immediately as the payment arrives it's made certain that the invoice
   has a lifetime of at least 2 hours left. This extension is done
   silently (except logged of course)
2. A cronjob designed to run hourly (to make sure it runs before an
   invoice extended by 2 hours expires) will extend it in steps of 24
   hours if it remains in pending state. If this happens a notification is
   sent, so the admin can make a decision on whether to cancel or not.

Do this by implementing the functionality generically in the
InvoiceManager so it can also be used by other payment methods in the
future, but Trustly being the only one it's implemented for right now.