Add support for generic managed bank accounts
This adds a generic support for a new type of payment method that is a
managed bank account. This is an account where we can download (or
otherwise get notified about) new transactions, and want them in our
systems. Support for actual providers are not included yet, and will be
done separately.

Incoming transactions can be matched against pending invoices using the
new payment references. If these match, it will be handled like any
other payment in the system.

They can also be matched against "bank matchers" which is a pattern
that other parts of the system can register for integration, for example
a creditcard provider can register a payout as going into a managed
account. This turns the payout into a two step operation, where the
first one is creating an accounting record and awaiting the payment, and
when the pattern is matched the record is automatically closed an filed.

Finally, all transactions that don't match either of the above can be
processed manually. This include both fuzzy matching of invoices (e.g.
payment reference is wrong) and creating manual accounting records for
manual transactions.

As part of this, the idea of "include bank information on the invoice"
is removed as a parameter on invoices. Instead, when the PDF is
generated the list of available invoice methods will be enumerated and
the first one with a bank information will have that included. If no
method includes bank information, it will be omitted from the invoice
(as will the payment reference).

A new scheduled job is added that sends a notification if there are
pending bank transactions in the system that have not been handled after
24 hours or more.

NOTE! As the bank information migrates into the database, it has to be
manually added to existing unmanaged bank trasfer providers, otherwise
no invoices at all will get bank information.