Remove old style bulk payments
This is replaced with the new style multi-registrations that have turned
out to be a lot closer to what the people using them actually wanted.

Internally the concept of bulk payments still exist, and the data models
remain unchanged.

The user visible interface for "bulk payment" is simply removed, and all
references to it in user-facing messages are replaced with
"multi-registration".