Update the Paypal access to use the new REST based API
The API we've been using has been deprecated for a while, and they've
now apparently started randomly dropping parts of it. So bite the bullet
and migrate to the new one. This is a bit cleaner for some things, lots
of things are missing, and transactions take several hours before they
show up in the reporting parts of it. Luckily, those parts are only used
in cronjobs running every few hours anyway...

There are new transaction codes for everything, so there are probably
some missed. There are also new and fancy inconsistencies in the
different ways that transaction texts can be reported, so there are most
likely some missed cases, but the common ones are covered.

Regular payment information is still using the PDT system (which is
independent of both APIs) and should not be affected. Only the fetching
of transactions, balances and issuing of refunds are.

NOTE: This requires reconfiguration of all paypal integrations. Instead
of api user/password/signature, an OAuth client and secret is used, and
the app itself has to be registered on each account. Instructions are in
the documentation.