Generalize frontend side forms
Reomve a lot of duplicates, and also make the frontend forms actually
use the bootstrap 4 templates that we have for everything else, instead
of using old hardcoded tables.

We already had generalized forms for the backend, but they include a
number of tings that are very backend-specific, so create a simplified
version for the frontend.