Introduce a shipment tracking system
This system allows for tracking individual shipments made primarily by
sponsors, but also by the conference organizers.

Each shipment it sent to an address, and multiple addresses can be
defined for a conference. Addresses can be made available on a
sponeorship level basis, meaning for example top tier sponsors can get
more options.

Each shipment is given a 5 digit random number. This is inserted into
the actual shipment address given to the sponsor. Hopefully this means
they will actually use said number, and thereby making it possible to
track the individual shipments.

A no-login-requierd very barebones interface is created for flagging
incoming shipments as delivered. In some cases this can be used directly
by hotel or partner receiving packages (thus does not require login),
and if not the partner can send information to the organizers who can
then update this interface, which triggers emails to the sponsors.