Ensure invoice processors are always run in a (sub)transaction
In case an excpetion occurrs in the invoice processor, weird things
could previously happen. In particular, when running from a notify
handler that was not wrapped in a transaction of it's own, partial
updates could happen.

This became obvious when the confreg processors started doing more work
in the processor, thereby increasing the likelihood of crashing.
Previously it would just do two simple updates, but with the addition of
ticket processing, it renders PDFs. PDF rendering is of course much more
likely to cause actual problems, and if this happens we don't want part
of the changes to go through and not other parts.

If this ends up running as part of an outer transaction, we will just
roll back the changes made as part of the invoice processor, using a
savepoint. The caller is then expected to check the result of the
function and take appropriate action, such as logging or sending an
email.