Add the concept of a public discount codes
This only exists in order to separate the public from non public codes
using separate tokens in the access token system. This way something
like an EARLYBIRD code can be flagged as public and be used to
dynamically update a website, while private codes are ont leaked in the
API call.