Make country optional and show most common countries
With country mandatory we see people picking something from the top of
the list, probably just because it's easy or because they don't want to
share their country.

To deal with the first, modify the country selection so that it first
shows the most common countries used for this conference, and only later
the full list of countries. The initial list of countries under "common
countries" can be specified at the conference configuration, and will then
gradually be replaced with the actual countries of the attendees.

To deal with the second, make the country field optional.