Make it possible to "unclaim" a sponsorship benefit
If a sponsor has claimed something in error, this makes it possible for
the administrator to undo that and move the benefit back to unclaimed
status. An email is of course generated both to the admins and to the
sponsor informing them of this.

In passing, fix some incorrect locations of </div> on sponsorship admin
page.