Enforce cfp workflow when sessions are set through voting page
Now that we have the workflow properly defined in the code, make sure
only the correct buttons show up in the dialog when changing the status
from the session voting page, and also enforce at server level that the
rules are not bypassed.

According to the docs we already did this, but it was never actually
enforced.