Introduce sponsor badge scanning as a benefit
This ads the ability to print a QR version of the attendee public token
(which is different from the regtoken -- used to pick up a disconnected
registration -- and the identity token which is used on tickets to scan
at checkin). Combined with this is a new sponsorship benefit to scan
badges.

Once claimed, that benefit gives access to the badge scanning system for
that sponsor, which can assign one or more of their attendees (typically
the booth staff) as scanners. These attendees will get access to a web
app similar to the one for checkins, and can use it to scan the public
tokens of attendee badges. For each such scan they can also make a note
attached to it.

The sponsorship administrators will then be able to view or download
(CSV) the information about those attendees scanned by their "team". The
information provided is name, company, country, and email address.

If badge scanning is enabled at the conference, each attendee get to
choose if they want to allow their badge to be scanned or not by
selecting it during the registration or updating it after the
registration is complete (similar to how photo consent works). They can
change their mind at any point, except if their badge has already been
scanned they can no longer opt-out (since for example the sponsor may
have taken a screenshot and we can't control that).