Make it possible to edit badge-scan-optin after locking reg
This field can only be changed in one direction already, so keeping it
unlocked makes sense. This allows attendees to opt-in to sharing their
data even after the event has started.

Make the code generic so we can turn on/off other fields for this as
well, should we want to.