Add support for "return to sender" on bank transfers
This allows the return of a bank transfer that does not belong to an
invoice. A regular refund is attached to an invoice and thus cannot be
used.

Implement this for TransferWise only at this point, since it's the only
bank transfer provider that actually supports sending money. We do this
by adding a generic payout framework to it, that we can also use for
other things in the future, and then use this one to track
return-payments.