Support spaces in payment references
Some people input payment references with spaces in them and since
that's always irrelevant to us, remove them in automatic matching.

Also update the matching of invoices to be able to show more than one
match, such as both the invoice number and the amount. Normally this
wouldn't happen, since it would've already been auto-matched, but as a
way to search things after the fact it can still be useful.