Remove TREASURER_EMAIL references from code, leaving in templates
TREASURER_EMAIL is supposed to be used in templates, but actively by the
code, but usage in the wrong way had snuck in. Remove that, and to deal
with that add a field to TransferWise configuration for
notification_receiver.

While at it, put a comment in the config file explaining what it's for.