Treat unmatched paypal transaction strings as empty ones
For historical reasons (pre-accounting-system, really) we treated
payments we didn't know what they were as errors, *except* if they were
sent with a completely empty string.

This changes them to instead be treated the same as the empty ones,
which means create an open accounting entry and let the operator deal
with it (but through the nicer interface, not through having to
hack the backend tables). It does this by separating out the "did not
match regexp" as a separate return value, and using that. Any other
errors in processing (such as actually finding the pattern with an
invoice number in it, but then not finding the invoice) is still treated
as an error like before.