Fix automatic notifications for reserve list talks
Previously if a reserve list talk was confirmed (by the speaker), the
notification was set to "last notified: confirmed", meaning that the
next time a manual notification run was made, the notification was sent
again.

Instead the "last notified" field should of course be set to the new
status of the session, whether confirmed or reserve-confirmed.

Bug spotted by ads.