Fix editing of cfp talks when a track has been closed
If a talk was submitted on track X, and track X was then subsequently
removed from the list of tracks available in the cfp, the speaker would
be forced to change the track of the talk in order to make any other
changes.

When this happens, now instead lock the submission so the track cannot
be changed at all (it can of course still be changed by an
administrator). Newly submitted talks can keep using the tracks that are
currently active for cfp.

Issue spotted by ads