Make custom jinja filters available in raw rendering
This is used for example when generating tweets for conference sponsors.
Previously the custom filters were only available when rendering actual
pages, but there is no reason to limit it to that.

In passing, add new jinja filter to apply a macro (used to apply a macro
inside a loop).