Fix scheduling of confreg_expire_additionaloptions
When moving to the built-in scheduler this job was changed to only
trigger in coordination with the waitlist. This was wrong, as it also
has to be triggered independent of that, once it's the additional option
that's about to sell out and not the conference itself.

So change it to run every 4 hours (checking for necessity each time of
course)