Automate handling of manually issued prepaid vouchers
We have 4 kinds of prepaid voucher generation, and 3 of them were
automatic - the ones where the sponsors buy vouchers or get them as a
benefit and t he ones where we generate a voucher manually with no
invoice.

This changes is so that the remaining case, when we generate vouchers
manually but invoice them, is now fully automated. Instead of manually
generating an incomplete voucher and then having to remember when to
share the actual voucher codes, piggyback on the system for sponsor
purchased vouchers. Create an invoice with the required details and once
that invoice is paid automatically generate the vouchers and an email to
the person buying them.

This breaks down the boundary between the confreg and the confsponsor
modules a bit more by reusing a confsponsor invoice processor, but there
are so many things where they are interdependent already that this makes
no realistic difference.