Include unused-but-available-rooms in schedule hash
Previously only rooms that had a session scheduled in them were passed
to the schedule template, making it difficult for to generically figure
out how to draw the schedule.

Since we now have the concepot of "rooms available on specific days",
add any rooms to the hash that is listed as available but has no
sessions. It will show up with an empty list of sessions of course, but
the availability of the room itself should help.

As existing conferences did not have the "available days" functionality,
there is no change in the hash for those conferences, so there should be
no issues with old templates.