Make "public visibility" a separate property on a sponsorship level
Previously if a level was just marked as "not available", it would still
be listed on the signup page, just not available for signup. By *also*
setting a level to "not publicly visible" it will now be removed
completely from this page, which makes it a lot easier to add things
temporarily for testing.