Redesign how attendee reports filters are generated
Instead of a long list of checkboxes and strange popouts, built them
using an "add filter" functionality and collect all the filters in a
table (backed by a json array). Bring this together with all the other
fields and submit as a single json request.

This also includes functionality to save and load report definitions in
the system, so that when a more complicated set of filters and fields
has been set it can be re-used.