Re-implement advanced attendee reports using SQL
The django-orm one was convenient in many ways, but often degraded into
loop-of-queries which wasn't great once we started getting to
conferences that had a fair number of attendees. And it also limited the
abilities to do filtering, since it's quite naive in it's understanding
of what SQL can do (and evenmore so in our "old" stable version).

So instead, create a trivial SQL generator (not a generic one, one just
for our case, which turns out not to be that complicated) that generates
native SQL and knows about both modern SQL and some PostgreSQL
extensions.

In doing this, add the ability to filter additional options based on
"has at least <n> of the options <x>,<y>,<z>", which can be used to
efficiently count attendees for trianing days at pgconfeu.

And finally, the queries for badges (and as such the JSON dump that is
supposed to have the exact same contents) are now hard-coded to in the
select part to be able to use decently advanced SQL. They still use the
same filtering functionality as the rest of the queries.