Add function to send TransferWise monthly statements to treasurer
TransferWise doesn't send statements automatically, and downloading them
is limited by the single-user-per-account-system. But since their API
allows downloading of a PDF statement between arbitrary dates, use that
to download the statements once per month and email them as an
attachment.

The function is optional and turned on/off on a per-paymentmethod basis.