Remove 250 character limit on tweet template
Since this is the *template* it makes no sense -- when for example
defining a macro that turns into no content at all in the output.

Instead, enforce the truncation of the tweet when it's generated. If it
truncates in a bad place, it can be fixed when moderating in the queue.