Store scanned badge directly witohut requiring Save button
Experiece from real-world use shows that it's common for sponsors to
scan a badge but not hit the save button, and then move on to do
something else, resulting in the badge scan not being stored.

Instead, store the scan immediately upon search (provided the attendee
allowed it, of course), and then update it laster if the sponsor writes
a note and hits save.