Fix confirmation for registrations off the waitlist
If a registration was first on the waitlist, and then the waitlist
registration was completed, and then the attendee went *back* to the
confirmation page (there would be no link to this page anymore, but if
it was stored in a bookmark on in a page that had not been refreshed
it's of course possible to get there), the waitlist entry would be
canceled and the attendee informed of that.

However, the actual registration would still not be canceled, since it
was confirmed, leading to much confusion.

Instead, if this happens, just redirect the user back to the
registration dashbaord without making any changes.