Implement incoming tweet polling in the trivial twitter client
If activated, this will poll for incoming tweets mentioning the
conference twitter account, and put them in a queue for processing
(which can mean either replying to them or discarding them as not
 needing a reply).

Interface is very simplistic, and the posting part of the system uses
the one already used for posting independent tweets.

The polling has to be enabled with a separate checkbox on the
integration page. It should normallyi not be more than one conference
with polling active for the same account, though in theory it could
work.