Change twitter timewindow to apply only to automated tweets
Now that there is a manual function for tweeting, it makes no sense to
restrict this one to the time window used for automatic tweets, thereby
for example making it impossible to reply to tweets outside this time.

Instead change it around so that this definition only applies to
automated tweets coming from the news or from confirmed sponsor
benefits.