Properly account for pending additional options in a lot of places
A pending additional option happens when somebody adds something (such
as training) to their registration after the registration is completed.
Once the invoice for the pending order has been paid, the option is
added to the registration, but until the invoice is paid it is in a
pending state. Unfortunately most places (dashboard, report and most
importantly the limit-number-of-registrations code) ignored those that
were in pending state. With this fix, they are now counted just like the
options added to a regular registration when it's in the invoiced state.

Pending orders can never go into "confirmed" state in these reports,
since at that point the option is added to the main registration record
and is counted there.

Thanks to Vik for some review and comments around the reporting queries.