Don't force sponosrship contract pdf to both exist and not
At some point the system started enforcing that a sponsorship contract
both had to be uploade din the first save, and also could not be
uploaded until it had been saved once, making it only possible to change
contracts and not create new ones.

This is an ugly fix removing half that check. It still requires a
contract to be created without a pdf and then edited to add the pdf. It
shuld be rewritten to use a better way of storage so this is not
required (we have other examples of that), but it works as a quick-fix.