Make sponsorships with zero cost work
For this, no invoice is generated. Instead they require a contract, and
any payment is handled completely outside the automated system. When the
operator has confirmed that payment has happened, verification is fully
manual.

Manual verification was always possible, but generating a zero cost
invoice could cause all sorts of issues elsewhere in the system.

In passing, change the step to confirm a sponsorsihp in the pre-existing
setup with a signed contract to use a POST instead of a GET, as should
be for operations that modify things.