Simplify InlineEncodedStorage and use bytea in table
Now that we are not using it as a django field anymore, simplify
InlineEncodedStorage to just natively do the two methods that we
actually need.

While at it, change it to use bytea internally instead of base64 encoded
in text, and add an appropriate unique constraint that was previously
missing.