Move speaker photos into main model with bytea
Instead of having a separate table with just the photos, move into the
main table. And instead of storing it base64 encoded, store it as a
proper bytea field.

Also rename the field from photofile to photo, to be more logical. Note
that this can BREAK templates overriding this field. The old name is
preserved on the model so that old templates for showing the page should
work, but not the form in the speaker profile.