Fix error reporting on failed twitter DM sending
If a message failed to send for some reason this was written as an error
message to the output, but the error flag was not set so this was
most likely never noticed.

This includes errors such as "can't send DM to somebody who is not
following us" (which will be annoying to get messages about, but they
should be reasonably few), but also important things like the twitter
integration being set up with incorrect permissions and *nothing*
working. It's much more important that those are not lost...