Add the idea of monitoring endpoints, and minimalistic git monitor
Create a namespace under /monitor/ that can be used for more monitoring
endpoints in the future, and a setting for MONITOR_SERVER_IPS to control
who is allowed to get information from them. Intended to be extended
with further points in the future.

For now, implement a "git" endpoint that returns information about the
current git position of this deployment (branch, tag if there is one,
and latest commit seen).