Make loading of qrencode and cairosvg optional
In particular, cairosvg is only used for the PNG format twitter cards,
so most functionality exists without it.

Make sure we import those modules conditionally only when needed and
handle the import exception. Also add an explicit attempt to load them
during system startup and write to the log if it fails, so it doesn't do
so silently.