Fix backend editing of sessions
The problem (exception when trying to edit) came from the fact that the
Speaker object was no longer pickle:able, meaning it could not be used
to generate the "concurrent editing protection" hash. This made it
impossible to edit any object that was *linked* to a Speaker -- but
there had to be an actual value. That is, editing sessions with no
speakers assigned worked fine because no Speaker object was ever
pickled, but when editing an object that had one or more Speaker records
linked, it would crash.

Fix that by making sure that the Speaker object converts its own photo
field from memoryview to bytes explicitly when being reduced, which
makes it pickle:able.