Fix call for papers list templates
When a user had no existing speaker profile *and* went to the "list of
open call for papers" (linked from the account home page), they would be
presented with a page that said "fill out speaker profile form below" but
there was no form. Instead, link the user to the speaker profile page
which does have a form.

That page, on the other hand, referred to "the form below", when the
actual form was above. So fix that too.

If the user had a speaker profile, or if they went directly to the cfp
for a specific conference (through the listing of *events* in the main
menu, just not through the listing of call for papers linked from the
profile), everything worked, which is why this went unnoticed for so
long. But with a completely new setup where all speakers are new
speakers, the likelihood of hitting the issue is of course much bigger.

Spotted by Dan Langille.